fix(VDP): extracted unsafe VDP primitives.
Moved VDP unsafe primitive functions from `vdp.h` to `vdp_unsafe.h`.
Also, added docstring for API reference of unsafe primitive functions.
